+++ cl_kern/sub.cl
@@ -14,11 +14,11 @@
 
 void kernel sub(global const double *S, global const double *I,
                 global const ushort *mask, global double *D,
-                const long convWidth, const long w, const long h,
+                const int convWidth, const int w, const int h,
                 const double convFactor, const double finalFactor) {
   const int id = get_global_id(0);
-  const long x = id % w;
-  const long y = id / w;
+  const int x = id % w;
+  const int y = id / w;
 
   int halfConvWidth = convWidth / 2;
   double d = 1e-30;
